+++ constants/type/modal.ts
@@ -5,12 +5,3 @@ export interface SelectProps {
   data: { index: number; label: string; value: string }[];
   onChange?: (e: { target: { value: string } }) => void;
 }
-
-export interface FilterProps {
-  modalType?: keyof typeof modal;
-  data: {
-    title: string;
-    type: { index: number; label: string; value: string }[];
-  }[];
-  onChange?: (e: { target: { value: string } }) => void;
-}
